GEODE-3301: Cargo installs now delete left over directories
This should hopefully fix Jenkins build issues
