+++ geode-assembly/src/test/java/org/apache/geode/session/tests/ContainerInstall.java
@@ -29,6 +29,7 @@ import javax.xml.transform.TransformerFactory;
 import javax.xml.transform.dom.DOMSource;
 import javax.xml.transform.stream.StreamResult;
 
+import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.logging.log4j.Logger;
 import org.codehaus.cargo.container.installer.Installer;
@@ -112,6 +113,9 @@ public abstract class ContainerInstall {
       String moduleName) throws IOException {
     this.connType = connType;
 
+    // Removes previous run stuff (modules, installs, etc.)
+    clearPreviousRuns();
+
     logger.info("Installing container from URL " + downloadURL);
 
     // Optional step to install the container from a URL pointing to its distribution
@@ -139,6 +143,23 @@ public abstract class ContainerInstall {
 
   public ServerContainer generateContainer(String containerDescriptors) throws IOException {
     return generateContainer(null, containerDescriptors);
+  }
+
+  /**
+   * Cleans up the installation by deleting the extracted module and downloaded installation folders
+   */
+  public void clearPreviousRuns() throws IOException {
+    File modulesFolder = new File(DEFAULT_MODULE_EXTRACTION_DIR);
+    File installsFolder = new File(DEFAULT_INSTALL_DIR);
+
+    // Remove default modules extraction from previous runs
+    if (modulesFolder.exists()) {
+      FileUtils.deleteDirectory(modulesFolder);
+    }
+    // Remove default installs from previous runs
+    if (installsFolder.exists()) {
+      FileUtils.deleteDirectory(installsFolder);
+    }
   }
 
   /**
@@ -216,30 +237,6 @@ public abstract class ContainerInstall {
    */
   public File getCacheXMLFile() {
     return new File(MODULE_PATH + "/conf/" + getConnectionType().getCacheXMLFileName());
-  }
-
-  /**
-   * Get the server life cycle class that should be used
-   *
-   * Generates the class based on whether the installation's connection type (@link #connType) is
-   * client server or peer to peer.
-   */
-  public String getServerLifeCycleListenerClass() {
-    String className = "org.apache.geode.modules.session.catalina.";
-    switch (connType) {
-      case PEER_TO_PEER:
-        className += "PeerToPeer";
-        break;
-      case CLIENT_SERVER:
-        className += "ClientServer";
-        break;
-      default:
-        throw new IllegalArgumentException(
-            "Bad connection type. Must be either PEER_TO_PEER or CLIENT_SERVER");
-    }
-
-    className += "CacheLifecycleListener";
-    return className;
   }
 
   /**
